+++ src/Entities/BuyerBalance.php
@@ -7,7 +7,7 @@ use \Hydra\Exchange\Interfaces\Entities\Pair as iPair;
 
 class BuyerBalance extends Abstracts\Balance
 {
-    public function __construct(int $primary, int $secondary)
+    public function __construct(float $primary, float $secondary)
     {
         return parent::__construct($primary, $secondary, Abstracts\Balance::OWNER_BUYER_TAKERER);
     }
