+++ src/Entities/SellOrder.php
@@ -8,7 +8,7 @@ use \Hydra\Exchange\Exceptions\Balance as BalanceException;
 
 class SellOrder extends Abstracts\Order
 {
-    public function __construct(iPair $pair, int $quantity, float $price, iBalance $balance, int $order = null)
+    public function __construct(iPair $pair, float $quantity, float $price, iBalance $balance, int $order = null)
     {
         if ($balance->getSecondary() < $quantity) {
             throw new BalanceException("insufficient funds (sell order) " . $balance->getPrimary() . " < " . $quantity . ")");
